# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "REST support for iBeer API",
    "description": "This is the REST UI support of iBeer API. Here you can perform REST operations on the API.",
    "contact": {
      "name": "Iago Paixao",
      "url": "https://github.com/iagopaixao/ibeer-service",
      "email": "euiagopaixao@gmail.com"
    },
    "license": {
      "name": "Apache License 2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/manufacturers": {
      "get": {
        "tags": [
          "manufacturer-resource"
        ],
        "operationId": "getAll",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageManufacturerResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "manufacturer-resource"
        ],
        "operationId": "update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManufacturerDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ManufacturerResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "manufacturer-resource"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManufacturerDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ManufacturerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/beers": {
      "get": {
        "tags": [
          "beer-resource"
        ],
        "operationId": "getAll_1",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageBeerResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "beer-resource"
        ],
        "operationId": "update_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeerDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BeerResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "beer-resource"
        ],
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeerDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BeerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/manufacturers/{id}": {
      "get": {
        "tags": [
          "manufacturer-resource"
        ],
        "operationId": "getOne",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ManufacturerResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "manufacturer-resource"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/beers/{id}": {
      "get": {
        "tags": [
          "beer-resource"
        ],
        "operationId": "getOne_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BeerResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "beer-resource"
        ],
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ManufacturerDTO": {
        "required": [
          "birthplace",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "birthplace": {
            "type": "string"
          }
        }
      },
      "BeerResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "style": {
            "type": "string"
          },
          "ibu": {
            "type": "integer",
            "format": "int32"
          },
          "abv": {
            "type": "number"
          },
          "milliliter": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number"
          },
          "manufacturer": {
            "type": "string"
          }
        }
      },
      "ManufacturerResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "birthplace": {
            "type": "string"
          },
          "beers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BeerResponse"
            }
          }
        }
      },
      "BeerDTO": {
        "required": [
          "abv",
          "ibu",
          "manufacturer",
          "milliliter",
          "name",
          "price",
          "style"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "ibu": {
            "maximum": 120,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "abv": {
            "type": "number"
          },
          "style": {
            "type": "string"
          },
          "price": {
            "minimum": 1,
            "type": "number"
          },
          "milliliter": {
            "type": "integer",
            "format": "int32"
          },
          "manufacturer": {
            "$ref": "#/components/schemas/ManufacturerDTO"
          }
        }
      },
      "PageManufacturerResponse": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManufacturerResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "paged": {
            "type": "boolean"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          }
        }
      },
      "PageBeerResponse": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BeerResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      }
    }
  }
}